refactor: relocate sites domain to global/link
Move AndroidAppLinkEndpoint, SharePageEndpoint, and UniversalLinkEndpoint from domain/sites to global/link as these are infrastructure-level link handling components, not business domain entities.
