+++ services/lifepuzzle-api/src/main/java/io/itmca/lifepuzzle/global/link/SharePageEndpoint.java
@@ -1,4 +1,4 @@
-package io.itmca.lifepuzzle.domain.sites;
+package io.itmca.lifepuzzle.global.link;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
